+++ R/utilities.R
@@ -986,7 +986,7 @@ GroupCorrelation <- function(
   grp.cors <- as.data.frame(x = grp.cors[which(x = !is.na(x = grp.cors))])
   grp.cors$gene_grp <- gene.grp[rownames(x = grp.cors)]
   colnames(x = grp.cors) <- c(paste0(var, "_cor"), "feature.grp")
-  object[[assay]][] <- grp.cors
+  object[[assay]] <- AddMetaData(object[[assay]], grp.cors)
   if (isTRUE(x = do.plot)) {
     print(GroupCorrelationPlot(
       object = object,