+++ R/utilities.R
@@ -986,7 +986,7 @@ GroupCorrelation <- function(
   grp.cors <- as.data.frame(x = grp.cors[which(x = !is.na(x = grp.cors))])
   grp.cors$gene_grp <- gene.grp[rownames(x = grp.cors)]
   colnames(x = grp.cors) <- c(paste0(var, "_cor"), "feature.grp")
-  object[[assay]][] <- grp.cors
+  object[[assay]] <- AddMetaData(object[[assay]], grp.cors)
   if (isTRUE(x = do.plot)) {
     print(GroupCorrelationPlot(
       object = object,
@@ -1171,11 +1171,31 @@ PercentageFeatureSet <- function(
   assay = NULL
 ) {
   assay <- assay %||% DefaultAssay(object = object)
-  if (!is.null(x = features) && !is.null(x = pattern)) {
-    warn(message = "Both pattern and features provided. Pattern is being ignored.")
+  if (!is.null(x = features))  {
+    if (!is.null(x = pattern)) {
+      warning(
+        paste(
+          "Both `features` and `pattern` were provided;",
+          "`pattern` will be ignored."
+        )
+      )
+    }
+
+    available_features <- Features(object[[assay]])
+    missing_features <- setdiff(features, available_features)
+    if (length(missing_features) > 0) {
+      warning(
+        paste(
+          "The following features are not found in the",
+          paste0("'", assay, "'"),
+          "assay:",
+          paste(paste0("'", missing_features, "'"), collapse = ", ")
+        )
+      )
+    }
   }
   percent.featureset <- list()
-  layers <- Layers(object = object, search = "counts")
+  layers <- Layers(object = object, assay = assay, search = "counts")
   for (i in seq_along(along.with = layers)) {
     layer <- layers[i]
     features.layer <- features %||% grep(
@@ -1185,6 +1205,7 @@ PercentageFeatureSet <- function(
     layer.data <- LayerData(object = object,
                             assay = assay,
                             layer = layer)
+    features.layer <- intersect(features.layer, rownames(layer.data))
     layer.sums <- colSums(x = layer.data[features.layer, , drop = FALSE])
     layer.perc <- layer.sums / object[[]][colnames(layer.data), paste0("nCount_", assay)] * 100
     percent.featureset[[i]] <- layer.perc