+++ R/generics.R
@@ -776,6 +776,17 @@ SCTResults <- function(object, ...) {
   UseMethod(generic = 'SCTResults', object = object)
 }
 
+#' Get the Pearson residuals from an sctransform-normalized dataset.
+#' 
+#' @param object An object
+#' @param ... Arguments passed to other methods (not used)
+#'
+#' @rdname FetchResiduals
+#' @export FetchResiduals
+#'
+FetchResiduals <- function(object, ...) {
+  UseMethod(generic = 'FetchResiduals', object = object)
+}
 
 #' @param value new data to set
 #'